refactor :: verify email package
