+++ feature/verify-email/src/main/java/team/retum/jobis/verify/email/navigation/VerifyEmailNavigation.kt
@@ -1,10 +1,10 @@
-package team.retum.jobis.verify_email.navigation
+package team.retum.jobis.verify.email.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import team.retum.common.enums.ResetPasswordNavigationArgumentType
-import team.retum.jobis.verify_email.ui.VerifyEmail
+import team.retum.jobis.verify.email.ui.VerifyEmail
 
 const val NAVIGATION_VERIFY_EMAIL = "verifyEmail"
